Revert "Extended client types"
This reverts commit c91942738224c84b95a427eff0496ce29dcd2d5e.
