+++ types/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@gbaranski/types",
-  "version": "3.1.4",
+  "version": "3.1.3",
   "description": "Types for controlhome app and website",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
