+++ types/src/index.ts
@@ -1,6 +1,5 @@
 export * from './alarmclock';
 export * from './watermixer';
 export * from './misc';
-export * from './device';
 export * from './client';
 export * from './firebase';
